Better D-Pad controlls.
Instead of accessing the dpad as an analog axis, this has now been
abstracted away to a userfriendly DPAD(LEFT) or DPAD(RIGHT) etc.

To always trigger when values are nonnull, just use DPAD(), for example
when you want to use the dpad to navigate your list and you want to
handle the different values in code rather than in config.